input validation and cleaning up messy code

+++ server/BuilderApi/Controllers/ExpenseCategories/CategoryController.cs
@@ -1,24 +1,30 @@
+using BuilderServices;
 using BuilderServices.ExpenseCategories.ExpenseCategoryService;
-using BuilderServices.ExpenseCategories.ExpenseCategoryService.Request;
 using BuilderServices.ExpenseCategories.ExpenseCategoryService.Responses;
 using BuilderServices.ExpenseCategories.ExpenseCategoryChartService;
 using BuilderServices.ExpenseCategories.ExpenseCategoryChartService.Requests;
+using BuilderServices.ExpenseCategories.ExpenseCategoryService.Requests;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
-namespace BuilderApi.Controllers;
+namespace BuilderApi.Controllers.ExpenseCategories;
 
 [ApiController]
 [Route("api/expenses/categories")]
 [Authorize]
 public class CategoryController(
     ExpenseCategoryService categoryService,
-    ExpenseCategoryChartService categoryChartService
+    ExpenseCategoryChartService categoryChartService,
+    ValidatorService validatorService
 ) : ControllerBase
 {
     [HttpGet]
     public async Task<IActionResult> GetExpenseCategories([FromQuery] GetExpenseCategoriesRequest request)
     {
+        var validationResult = await validatorService.ValidateAsync(request);
+        if (!validationResult.IsValid)
+            return BadRequest(validationResult.Errors);
+        
         var categories = await categoryService.GetExpenseCategoriesAsync(request.Active).ConfigureAwait(false);
 
         return Ok(categories);
@@ -27,7 +33,9 @@ public class CategoryController(
     [HttpPost("create")]
     public async Task<IActionResult> CreateExpenseCategory([FromBody] CreateExpenseCategoryRequest request)
     {
-        // TODO: validate request
+        var validationResult = await validatorService.ValidateAsync(request);
+        if (!validationResult.IsValid)
+            return BadRequest(validationResult.Errors);
 
         var isCreated = await categoryService.CreateExpenseCategoryAsync(request.CategoryName).ConfigureAwait(false);
 
@@ -40,6 +48,10 @@ public class CategoryController(
     [HttpPatch("{id:int}/update/active")]
     public async Task<IActionResult> SetExpenseCategoryActiveStatus([FromBody] SetExpenseCategoryActiveStatusRequest request, int id)
     {
+        var validationResult = await validatorService.ValidateAsync(request);
+        if (!validationResult.IsValid)
+            return BadRequest(validationResult.Errors);
+        
         await categoryService.SetExpenseCategoryActiveStatusAsync(id, request.Active).ConfigureAwait(false);
 
         return Ok(new SetExpenseCategoryActiveStatusResponse
@@ -51,6 +63,10 @@ public class CategoryController(
     [HttpGet("totalSpent")]
     public async Task<IActionResult> GetCategoryTotalSpentByRangeAsync([FromQuery] CategoryTotalSpentRequest request)
     {
+        var validationResult = await validatorService.ValidateAsync(request);
+        if (!validationResult.IsValid)
+            return BadRequest(validationResult.Errors);
+        
         var categories = await categoryChartService.GetCategoryTotalSpentByRangeAsync(request.RangeOption).ConfigureAwait(false);
 
         return Ok(categories);
@@ -70,6 +86,10 @@ public class CategoryController(
     [HttpPatch("update/name")]
     public async Task<IActionResult> UpdateCategoryName(UpdateCategoryNameRequest request)
     {
+        var validationResult = await validatorService.ValidateAsync(request);
+        if (!validationResult.IsValid)
+            return BadRequest(validationResult.Errors);
+        
         await categoryService.UpdateCategoryNameAsync(request.CategoryId, request.NewCategoryName).ConfigureAwait(false);
 
         return Ok(new UpdateCategoryNameResponse
@@ -81,6 +101,9 @@ public class CategoryController(
     [HttpDelete("{id:int}/delete")]
     public async Task<IActionResult> DeleteExpenseCategory(int id)
     {
+        if (id <= 0)
+            return BadRequest("Category id must be greater than 0");
+        
         await categoryService.DeleteExpenseCategoryAsync(id).ConfigureAwait(false);
 
         return Ok(new DeleteExpenseCategoryResponse
@@ -90,14 +113,13 @@ public class CategoryController(
     }
 
     [HttpGet("avg")]
-    public async Task<IActionResult> GetAvgSpentForCategories([FromQuery] CategoryAvgSpentRequest? request)
+    public async Task<IActionResult> GetAvgSpentForCategories([FromQuery] CategoryAvgSpentRequest request)
     {
-        var year = request?.Year ?? DateTime.Today.Year;
-        if (year is < 1000 or > 9999)
-        {
-            return BadRequest("Invalid year.");
-        }
-        var result = await categoryChartService.GetAvgSpentForCategoriesAsync(year).ConfigureAwait(false);
+        var validationResult = await validatorService.ValidateAsync(request);
+        if (!validationResult.IsValid)
+            return BadRequest(validationResult.Errors);
+        
+        var result = await categoryChartService.GetAvgSpentForCategoriesAsync(request.Year).ConfigureAwait(false);
 
         return Ok(result);
     }
